+++ SVE/LightManager.h
@@ -16,8 +16,8 @@ public:
     LightManager();
     ~LightManager();
 
-    void setLight(std::shared_ptr<LightNode> light, uint16_t index);
-    std::shared_ptr<LightNode> getLight(uint16_t index) const;
+    void setLight(std::shared_ptr<LightNode> light, uint32_t index);
+    std::shared_ptr<LightNode> getLight(uint32_t index) const;
     size_t getLightCount() const;
 
     VulkanShadowImage* getVulkanShadowImage();
